@@ -1,7 +1,7 @@
 import Alert from "./Alert";
-import ReactDOM from "react-dom";
 import userEvent from "@testing-library/user-event";
-import { render } from "@testing-library/react";
+import { act, render } from "@testing-library/react";
+import { createRoot } from "react-dom/client";
 
 function buildComponent(props = {}) {
   props = { children: <div>test</div>, variant: "info", ...props };
@@ -9,8 +9,9 @@ function buildComponent(props = {}) {
 }
 
 it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  act(() => root.render(buildComponent()));
 });
 
 it("renders alert-<variant> className", () => {
