@@ -1,14 +1,15 @@
 import FieldError from "./FieldError";
-import ReactDOM from "react-dom";
-import { render, screen } from "@testing-library/react";
+import { act, render, screen } from "@testing-library/react";
+import { createRoot } from "react-dom/client";
 
 function buildComponent(props = {}) {
   return <FieldError {...props} />;
 }
 
 it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  act(() => root.render(buildComponent()));
 });
 
 describe("when error prop is given", () => {
