@@ -1,11 +1,12 @@
 import Footer from "./Footer";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 function buildComponent(props = {}) {
   return <Footer {...props} />;
 }
 
 it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  root.render(buildComponent());
 });
