@@ -1,7 +1,7 @@
-import ReactDOM from "react-dom";
 import RecipeTitle from "./RecipeTitle";
+import { act, render, screen } from "@testing-library/react";
+import { createRoot } from "react-dom/client";
 import { MemoryRouter } from "react-router-dom";
-import { render, screen } from "@testing-library/react";
 
 function buildComponent(props = {}) {
   props = {
@@ -20,8 +20,9 @@ function buildComponent(props = {}) {
 }
 
 it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  act(() => root.render(buildComponent()));
 });
 
 it("renders data.title prop", () => {
