@@ -1,12 +1,15 @@
 import App from "./App";
 import AuthnProvider from "./providers/AuthnProvider";
 import React from "react";
-import ReactDOM from "react-dom";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
+import { createRoot } from "react-dom/client";
 import { HelmetProvider } from "react-helmet-async";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <React.StrictMode>
     <HelmetProvider>
       <BrowserRouter>
@@ -15,8 +18,7 @@ ReactDOM.render(
         </AuthnProvider>
       </BrowserRouter>
     </HelmetProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
