@@ -4,11 +4,11 @@ jest.mock("../hooks/useAuthn", () => ({
 }));
 
 import Login from "./Login";
-import ReactDOM from "react-dom";
 import useAuthn from "../hooks/useAuthn";
+import { act, render, waitFor } from "@testing-library/react";
+import { createRoot } from "react-dom/client";
 import { HelmetProvider } from "react-helmet-async";
 import { MemoryRouter } from "react-router-dom";
-import { render, waitFor } from "@testing-library/react";
 
 function buildComponent() {
   return (
@@ -25,8 +25,9 @@ beforeEach(() => {
 });
 
 it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  act(() => root.render(buildComponent()));
 });
 
 it("renders the correct <title>", async () => {
