@@ -3,10 +3,10 @@ jest.mock("../hooks/useApi", () => ({
   default: jest.fn(),
 }));
 
-import ReactDOM from "react-dom";
 import SignupConfirmation from "./SignupConfirmation";
 import useApi from "../hooks/useApi";
-import { act, render, waitFor } from "@testing-library/react";
+import { act, render, screen, waitFor } from "@testing-library/react";
+import { createRoot } from "react-dom/client";
 import { HelmetProvider } from "react-helmet-async";
 import { MemoryRouter } from "react-router-dom";
 
@@ -27,29 +27,23 @@ beforeEach(() => {
   useApi.mockReturnValue({ post });
 });
 
-it("renders successfully", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(buildComponent(), div);
+it("renders successfully", async () => {
+  const container = document.createElement("div");
+  const root = createRoot(container);
+  await act(async () => root.render(buildComponent()));
 });
 
 it("renders the correct <title>", async () => {
   post.mockResolvedValue({ message: "test message" });
-
-  act(() => {
-    render(buildComponent());
-  });
-
+  await act(() => render(buildComponent()));
   await waitFor(() => expect(document.title).toContain("Signup Confirmation"));
 });
 
 describe("while sign up confirmation is pending", () => {
   it("renders a pending message", async () => {
     post.mockReturnValue(new Promise(() => {}));
-
-    act(() => {
-      const { queryByText } = render(buildComponent());
-      expect(queryByText("Confirming your signup...")).toBeTruthy();
-    });
+    await act(() => render(buildComponent()));
+    expect(screen.queryByText("Confirming your signup...")).toBeTruthy();
   });
 });
 
@@ -57,16 +51,13 @@ describe("when sign up confirmation fails", () => {
   it("renders an error message", async () => {
     const message = "test error";
     post.mockResolvedValue({ isError: true, message });
-    let queryByTestId, queryByText;
-
-    await act(async () => {
-      const container = render(buildComponent());
-      queryByTestId = container.queryByTestId;
-      queryByText = container.queryByText;
-    });
-
-    expect(queryByTestId("signup-confirmation-message-error")).toBeTruthy();
-    expect(queryByText(message)).toBeTruthy();
+    await act(() => render(buildComponent()));
+    await waitFor(() =>
+      expect(
+        screen.queryByTestId("signup-confirmation-message-error")
+      ).toBeTruthy()
+    );
+    expect(screen.queryByText(message)).toBeTruthy();
   });
 });
 
@@ -74,15 +65,12 @@ describe("when sign up confirmation is successful", () => {
   it("renders a success message", async () => {
     const message = "test success";
     post.mockResolvedValue({ message });
-    let queryByTestId, queryByText;
-
-    await act(async () => {
-      const container = render(buildComponent());
-      queryByTestId = container.queryByTestId;
-      queryByText = container.queryByText;
-    });
-
-    expect(queryByTestId("signup-confirmation-message-success")).toBeTruthy();
-    expect(queryByText(message)).toBeTruthy();
+    await act(() => render(buildComponent()));
+    await waitFor(() =>
+      expect(
+        screen.queryByTestId("signup-confirmation-message-success")
+      ).toBeTruthy()
+    );
+    expect(screen.queryByText(message)).toBeTruthy();
   });
 });
